Implements fireSpawnEvent
I realized that AMBSpawner doesn't need to be subclassed if it gets directly passed objects that wish to be spawned (e.g. a patient). I updated the init for Spawner so that it receives an array of objects. Multiple objects will be randomly chosen from every time a spawn event occurs.

Implemented a skeleton for loadSharedAssets that I want to use in the future.

TODO: test the spawner, and refactor the addToChild code to call a method that deals with adding it to the scene.

issue #4