WIP handleInput; implements manual lane change Holding down left or right moves the player at a constant speed; releasing the key "snaps" the player into a lane. Perfect!
TODO: Refactor the actual lane change code in authorizeMoveEvent and fold it into changeLanes, with a shouldSnap boolean to control behaviour. We still need to test for a valid target point when doing a manual lane change, so the only difference will be whether or not the action should snap to a lane at the end.